Fix thumb not being drawn on vertical seekbar
In Marshmallow the enabled thumb was not being drawn
due to a bug where the thumb is not being rotated and
translated like the rest of the seekbar.  This patch
implements a bit of a hack to get the thumb to be
displayed on devices running Android Marshmallow by
manually rendering the enabled thumb.